pyproject.toml
[project]
name = "multi-agent-clander-email-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair==5.3.0",
    "colorlog>=6.10.1",
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "langchain>=1.2.10",
    "langchain-core>=1.2.12",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "langgraph-checkpoint>=4.0.0",
    "numpy>=2.4.2",
    "openai>=2.21.0",
    "pandas==2.2.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "streamlit==1.54.0",
    "streamlit-extras>=0.1.5",
]
